refactor(core): move cross-layer exceptions to core/exceptions.py
Move StreamChunkTimeout, ScrapeFailed, and VaultWriteFailed to the
innermost layer (core) to eliminate layer violations:

- core→integrations: _scraping.py and vault_add.py imported from
  integrations for ScrapeFailed/VaultWriteFailed
- adapters→nats: _shared_streaming_state.py imported StreamChunkTimeout
  from nats layer

Exception classes are structural contracts — they should live in the
innermost layer that needs to raise or catch them.

Fixes P1 Architecture #3-4 from code quality audit.
