+++ tests/core/processors/test_vault_add_post.py
@@ -9,9 +9,10 @@ from unittest.mock import AsyncMock, MagicMock
 
 from lyra.core.auth.trust import TrustLevel
 from lyra.core.commands.command_parser import CommandContext
+from lyra.core.exceptions import VaultWriteFailed
 from lyra.core.messaging.message import InboundMessage, Response
 from lyra.core.processors.vault_add import VaultAddProcessor
-from lyra.integrations.base import SessionTools, VaultWriteFailed
+from lyra.integrations.base import SessionTools
 
 
 def make_tools(
