+++ tests/core/processors/test_vault_add_pre.py
@@ -8,10 +8,11 @@ from unittest.mock import AsyncMock, MagicMock
 
 from lyra.core.auth.trust import TrustLevel
 from lyra.core.commands.command_parser import CommandContext
+from lyra.core.exceptions import ScrapeFailed
 from lyra.core.messaging.message import InboundMessage
 from lyra.core.processors._scraping import _SAFE_SCRAPE_MAX_CHARS
 from lyra.core.processors.vault_add import VaultAddProcessor
-from lyra.integrations.base import ScrapeFailed, SessionTools
+from lyra.integrations.base import SessionTools
 
 
 def make_tools(
